+++ src/core/parser/CodefendParser.ts
@@ -1,8 +1,8 @@
 import {
   ICodefendOptions,
   ICodefendRegexListOption,
-} from "../options/ICodeDefendOptions";
-import { ICodefendParser, ICodefendParserWord } from "./ICodeDefendParser";
+} from "../options/ICodefendOptions";
+import { ICodefendParser, ICodefendParserWord } from "./ICodefendParser";
 
 export class CodefendParser implements ICodefendParser {
   options: ICodefendOptions | undefined;
@@ -18,7 +18,7 @@ export class CodefendParser implements ICodefendParser {
   parse(code: string, regexList?: ICodefendRegexListOption[]) {
     regexList = regexList ?? this.options?.regexList;
     if (!regexList) {
-      throw new Error("Code Defender: regexList required for parser");
+      throw new Error("Codefend: regexList required for parser");
     }
     const words: ICodefendParserWord[] = [];
     let match;
