+++ src/core/replacer/ICodefendReplacer.ts
